# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  types.FileUploadError:
    description: File upload error response payload
    properties:
      error:
        example: Invalid file
        type: string
      error_code:
        example: FILE_TOO_LARGE
        type: string
      message:
        example: File size exceeds maximum limit of 5MB
        type: string
    type: object
...
  types.UploadResponse:
    description: File upload response payload
    properties:
      avatar_url:
        example: /api/user/12345/avatar
        type: string
      message:
        example: Avatar uploaded successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    A comprehensive API for connecting sports enthusiasts, managing profiles, and facilitating sports-related activities.
    This API provides authentication, user profile management, search functionality, and file upload capabilities.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sport Connect API
  version: "1.0"
paths:
...
  /upload/avatar:
    delete:
      consumes:
      - application/json
      description: Remove the current user's custom avatar image, reverting to default
        avatar
      produces:
      - application/json
      responses:
        "200":
          description: Avatar deleted successfully
          schema:
            $ref: '#/definitions/types.UploadResponse'
        "400":
          description: No avatar to delete
          schema:
            $ref: '#/definitions/types.FileUploadError'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/types.FileUploadError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/types.FileUploadError'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/types.FileUploadError'
      security:
      - BearerAuth: []
      summary: Delete user avatar
      tags:
      - Upload
    post:
      consumes:
      - multipart/form-data
      description: Upload a new avatar image for the authenticated user. Accepts JPEG,
        PNG, GIF, and WebP formats. Maximum file size is 5MB.
      parameters:
      - description: Avatar image file (JPEG, PNG, GIF, WebP, max 5MB)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Avatar uploaded successfully
          schema:
            $ref: '#/definitions/types.UploadResponse'
        "400":
          description: No file uploaded or invalid file format
          schema:
            $ref: '#/definitions/types.FileUploadError'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/types.FileUploadError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/types.FileUploadError'
        "500":
          description: File processing or database error
          schema:
            $ref: '#/definitions/types.FileUploadError'
      security:
      - BearerAuth: []
      summary: Upload user avatar
      tags:
      - Upload
  /user/{userId}/avatar:
    get:
      consumes:
      - application/json
      description: Retrieve a user's avatar image by user ID. Returns custom uploaded
        avatar or redirects to default avatar
      parameters:
      - description: User ID
        example: 12345
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - image/jpeg
      - image/png
      - image/gif
      - image/webp
      responses:
        "200":
          description: User's avatar image
          schema:
            type: string
        "302":
          description: Redirect to default avatar
          schema:
            type: string
        "400":
          description: User ID is required
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: No avatar found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get user avatar
      tags:
      - Upload
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...